+++ airflow/dags_dinamicos/config_universidades_c.yaml
@@ -1,0 +1,4 @@
+dag_id: 'universidades_c'
+schedule_interval: '@hourly'
+catchup: False
+input: 2
